- packages and project renaming preparing for repo renaming
