+++ pom.xml
@@ -14,8 +14,8 @@
     <!--endregion-->
 
     <!-- region Your particular project data -->
-    <groupId>org.white_sdev.template.logger_db_runnable_template</groupId>
-    <artifactId>logger-db-runnable-template</artifactId>
+    <groupId>org.white_sdev.templates</groupId>
+    <artifactId>white_template</artifactId>
     <version>0.1-SNAPSHOT</version>
     <name>${project.groupId}:${project.artifactId}</name>
     <description>Description of my application</description>
@@ -27,7 +27,7 @@
         <!-- region Configure Java version & Main full package and class name -->
         <!-- WARNING: add release scope if 1.9 and older versions are used at build/plugins/plugin/[maven-compiler-plugin]/configuration-->
         <java.version>17</java.version>
-        <main.class>org.white_sdev.template.logger_db_runnable_template.LoggerDbRunnableTemplate</main.class>
+        <main.class>org.white_sdev.templates.white_template.LoggerDbRunnableTemplate</main.class>
         <!--endregion-->
         <!-- region General configurations of the project -->
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
